Bug: 97261 Ajuste na consolidação de frequência mensal, alterando o DTO respectivo para utilizar a classe padrão para definição do percentual de frequência. Alteração do tipo decimal para double do percentual de acordo com a classe padrão de frequência aluno.

+++ src/SME.SGP.Dominio/Entidades/ConsolidacaoFrequenciaAlunoMensal.cs
@@ -6,7 +6,7 @@
         public long TurmaId { get; set; }
         public string AlunoCodigo { get; set; }
         public int Mes { get; set; }
-        public decimal Percentual { get; set; }
+        public double Percentual { get; set; }
         public int QuantidadeAulas { get; set; }
         public int QuantidadeAusencias { get; set; }
         public int QuantidadeCompensacoes { get; set; }
